requirements.txt
telebot==0.0.5
python-dotenv
requests==2.32.3
pytube==15.0.0
yt-dlp==2024.8.6